pointset method flattened
Before the points method would return a list of pairs of points since each call to the square root function returns two square roots, so the point maker would return two points with the same abscissa or x value. The javascript would rather have it flattened, so we change the points method so that we don't have to reformat it later or something garbage like that.