GeneralField
Worked on the General Field and Element classes. The basic arithmetic
is tested over GF_5. Now, the next step is to check over the other
residues mod 8, and then to test the higher arithmetic like the square
roots and exponents. Once that’s done, fork so EllipticCurve.py is
defined over GeneralField.py.