Debuggin
Old mistakes were made again, e.g. initializers should not return a
value. Further, I misused the raise except clauses in the EllipticCurve
definition, so I had to rewrite those. THOSE ARE ALSO STILL WRONG
because the ELSE is outside of the try. So that’s all wrong. Fix that
later when we have time.