Fixing Class Relationships
Made it so that the Element Class feeds integer values to the GeneralField class' methods instead of Elements. This is I think the source of a lot of the errors...
Made more use of the of method. Makes the code a bit more concise.
Fixed operator function name. ** is implemented using __pow__ not __exp__.